mindmap/requirements.txt
pandas
numpy
scipy
plotly
openai == 0.27.9
tiktoken
langchain
python-dotenv